src/serve/requirements.txt
Flask
google-cloud-storage
joblib
scikit-learn
gcsfs
python-dotenv